Add files via upload
The script is created by Anand B Pillai (all credits are to him)

This program is an advanced directory synchronisation and update tool. It can be used to update content between two directories, synchronise them, or just report the difference in content between them. It uses the syntax of the 'diff' program in printing the difference.

THE AIM OF THIS PROJECT IS TO MODIFY THE SCRIPT TO MACE IT SYNCHRONISE TWO FOLDERS CONTINUOUSLY. SO IF THERE ARE ANY CHANGES TAKEN IN THE FIRST FOLDER, IT AUTOMATICALLY MIRRORING IT IN TO THE SECOND FOLDER, WHETHER IT IS ADDING THE NEW FILE/FOLDER OR DELETING FILE/FOLDER.

*Project is open to everyone!!!